[IOSAS] Test new patching functions
UTMT-CLI will now externalize textures and replace gml scripts on the fly. This removes the need for an xdelta patch. As a result, iosas now uses the dotnet runtime.

Scripts are extracted from the data.win file and modified with python, so only relevant and modified code is publicly distributed.